+++ evalml/tests/automl_tests/test_auto_classification_search.py
@@ -8,11 +8,13 @@ from sklearn.model_selection import StratifiedKFold, TimeSeriesSplit
 
 from evalml import AutoClassificationSearch
 from evalml.automl.pipeline_search_plots import SearchIterationPlot
+from evalml.exceptions import ObjectiveNotFoundError
 from evalml.model_family import ModelFamily
 from evalml.objectives import (
     FraudCost,
     Precision,
     PrecisionMicro,
+    Recall,
     get_objective,
     get_objectives
 )
@@ -43,6 +45,8 @@ def test_init(X_y):
 
     assert isinstance(automl.best_pipeline, PipelineBase)
     assert isinstance(automl.get_pipeline(0), PipelineBase)
+    with pytest.raises(RuntimeError, match="Pipeline not found"):
+        automl.get_pipeline(1000)
 
     automl.describe_pipeline(0)
 
@@ -100,6 +104,19 @@ def test_specify_objective(X_y):
     assert automl.best_pipeline.threshold is not None
 
 
+def test_recall_error(X_y):
+    X, y = X_y
+    error_msg = 'Could not find the specified objective.'
+    with pytest.raises(ObjectiveNotFoundError, match=error_msg):
+        AutoClassificationSearch(objective='recall', max_pipelines=1)
+
+
+def test_recall_object(X_y):
+    X, y = X_y
+    automl = AutoClassificationSearch(objective=Recall(), max_pipelines=1)
+    automl.search(X, y)
+
+
 def test_binary_auto(X_y):
     X, y = X_y
     automl = AutoClassificationSearch(objective="log_loss_binary", multiclass=False, max_pipelines=5)
@@ -111,7 +128,7 @@ def test_binary_auto(X_y):
 
 def test_multi_error(X_y_multi):
     X, y = X_y_multi
-    error_automls = [AutoClassificationSearch(objective='recall'), AutoClassificationSearch(objective='recall_micro', additional_objectives=['recall'], multiclass=True)]
+    error_automls = [AutoClassificationSearch(objective='precision'), AutoClassificationSearch(objective='precision_micro', additional_objectives=['precision'], multiclass=True)]
     error_msg = 'not compatible with a multiclass problem.'
     for automl in error_automls:
         with pytest.raises(ValueError, match=error_msg):
@@ -120,7 +137,7 @@ def test_multi_error(X_y_multi):
 
 def test_multi_auto(X_y_multi):
     X, y = X_y_multi
-    automl = AutoClassificationSearch(objective="recall_micro", multiclass=True, max_pipelines=5)
+    automl = AutoClassificationSearch(objective="f1_micro", multiclass=True, max_pipelines=5)
     automl.search(X, y)
     y_pred = automl.best_pipeline.predict(X)
     assert len(np.unique(y_pred)) == 3
@@ -146,10 +163,10 @@ def test_multi_objective(X_y_multi):
     automl = AutoClassificationSearch(objective="log_loss_multi")
     assert automl.problem_type == ProblemTypes.MULTICLASS
 
-    automl = AutoClassificationSearch(objective='recall_micro')
+    automl = AutoClassificationSearch(objective='auc_micro')
     assert automl.problem_type == ProblemTypes.MULTICLASS
 
-    automl = AutoClassificationSearch(objective='recall')
+    automl = AutoClassificationSearch(objective='auc')
     assert automl.problem_type == ProblemTypes.BINARY
 
     automl = AutoClassificationSearch(multiclass=True)
@@ -161,7 +178,7 @@ def test_multi_objective(X_y_multi):
 
 def test_categorical_classification(X_y_categorical_classification):
     X, y = X_y_categorical_classification
-    automl = AutoClassificationSearch(objective="recall", max_pipelines=5, multiclass=False)
+    automl = AutoClassificationSearch(objective="precision", max_pipelines=5, multiclass=False)
     automl.search(X, y)
     assert not automl.rankings['score'].isnull().all()
     assert not automl.get_pipeline(0).feature_importances.isnull().all().all()
@@ -237,7 +254,7 @@ def test_additional_objectives(X_y):
 def test_optimizable_threshold_enabled(mock_fit, mock_score, mock_predict_proba, mock_optimize_threshold, X_y, caplog):
     mock_optimize_threshold.return_value = 0.8
     X, y = X_y
-    automl = AutoClassificationSearch(objective='recall', max_pipelines=1, optimize_thresholds=True)
+    automl = AutoClassificationSearch(objective='precision', max_pipelines=1, optimize_thresholds=True)
     mock_score.return_value = {automl.objective.name: 1.0}
     automl.search(X, y)
     mock_fit.assert_called()
@@ -258,7 +275,7 @@ def test_optimizable_threshold_enabled(mock_fit, mock_score, mock_predict_proba,
 def test_optimizable_threshold_disabled(mock_fit, mock_score, mock_predict_proba, mock_optimize_threshold, X_y):
     mock_optimize_threshold.return_value = 0.8
     X, y = X_y
-    automl = AutoClassificationSearch(objective='recall', max_pipelines=1, optimize_thresholds=False)
+    automl = AutoClassificationSearch(objective='precision', max_pipelines=1, optimize_thresholds=False)
     mock_score.return_value = {automl.objective.name: 1.0}
     automl.search(X, y)
     mock_fit.assert_called()