+++ evalml/tests/pipeline_tests/classification_pipeline_tests/test_binary_classification.py
@@ -11,16 +11,16 @@ import pytest
 @patch('evalml.pipelines.PipelineBase.fit')
 def test_binary_classification_pipeline_predict(mock_fit, mock_transform,
                                                 mock_predict, mock_predict_proba,
-                                                mock_obj_decision, X_y, dummy_binary_pipeline):
+                                                mock_obj_decision, X_y, dummy_binary_pipeline_class):
     X, y = X_y
-    binary_pipeline = dummy_binary_pipeline
+    binary_pipeline = dummy_binary_pipeline_class(parameters={})
     # test no objective passed and no custom threshold uses underlying estimator's predict method
     binary_pipeline.predict(X)
     mock_predict.assert_called()
     mock_predict.reset_mock()
 
     # test objective passed but no custom threshold uses underlying estimator's predict method
-    binary_pipeline.predict(X, 'recall')
+    binary_pipeline.predict(X, 'precision')
     mock_predict.assert_called()
     mock_predict.reset_mock()
 
@@ -46,16 +46,16 @@ def test_binary_classification_pipeline_predict(mock_fit, mock_transform,
     mock_predict_proba.reset_mock()
     mock_predict_proba.return_value = np.array([[0.1, 0.2], [0.1, 0.2]])
     binary_pipeline.threshold = 0.6
-    binary_pipeline.predict(X, 'recall')
+    binary_pipeline.predict(X, 'precision')
     mock_predict.assert_not_called()
     mock_predict_proba.assert_called()
     mock_obj_decision.assert_called()
 
 
 @patch('evalml.pipelines.PipelineBase._transform')
-def test_binary_predict_pipeline_objective_mismatch(mock_transform, X_y, dummy_binary_pipeline):
+def test_binary_predict_pipeline_objective_mismatch(mock_transform, X_y, dummy_binary_pipeline_class):
     X, y = X_y
-    binary_pipeline = dummy_binary_pipeline
+    binary_pipeline = dummy_binary_pipeline_class(parameters={})
     with pytest.raises(ValueError, match="You can only use a binary classification objective to make predictions for a binary classification pipeline."):
-        binary_pipeline.predict(X, "recall_micro")
+        binary_pipeline.predict(X, "precision_micro")
     mock_transform.assert_called()